+++ vss_cli/const.py
@@ -5,7 +5,7 @@ import pkg_resources
 
 PACKAGE_NAME = "vss_cli"
 
-__version__ = "2022.3.1"
+__version__ = "2022.4.0"
 
 
 DEFAULT_TIMEOUT = 30